+++ src/assets/images/partners/import.ts
@@ -1,0 +1,6 @@
+import agro from "./agro.png";
+import flexa from "./flexa.png";
+import ht from "./hellermanntyton.png";
+import ivm from "./ivm.png";
+
+export { agro, flexa, ht, ivm };