+++ src/pages/Contact/map.tsx
@@ -6,7 +6,8 @@ const Map = () => {
   return (
     <Col md={6}>
       <h2 className="mb-4">{t("contact.location")}</h2>
-      <div style={{ height: "300px", background: "#eee" }}>
+
+      <div style={{ height: "427px"}}> {/*manual height otherwise wont fill*/}
         <iframe
           className="w-100 h-100"
           src="https://maps.google.com/maps?output=embed&amp;q=nef%2022%20atak%C3%B6y&amp;z=10&amp;t=m"