+++ src/pages/NotFound.tsx
@@ -1,23 +1,24 @@
 import { useTranslation } from "react-i18next";
 import { getLanguageFromRoute } from "@/locales/i18n";
 import Footer from "@/components/Footer";
+import { Container, Button } from "react-bootstrap";
 
 const NotFound = () => {
   const { t } = useTranslation();
   const currentLanguage = getLanguageFromRoute();
+
   return (
-    <>
-      <div className="d-flex justify-content-center vh-100">
-        <div className="text-center">
-          <h1 className="display-1">404</h1>
-          <p className="lead">{t("notFound.message")}</p>
-          <a href={`#/${currentLanguage}/`} className="btn btn-primary">
-            {t("notFound.goHome")}
-          </a>
-        </div>
+    // To fix the content to the top and footer to the bottom
+    <Container className="d-flex flex-column min-vh-100">
+      <div className="flex-grow-1 d-flex flex-column align-items-center text-center">
+        <h1 className="display-1">404</h1>
+        <p className="lead">{t("notFound.message")}</p>
+        <Button href={`#/${currentLanguage}/`} variant="primary">
+          {t("notFound.goHome")}
+        </Button>
       </div>
       <Footer />
-    </>
+    </Container>
   );
 };