+++ src/pages/home/Home.tsx
@@ -11,6 +11,7 @@ const Home = () => {
       <Section />
       <Catalog />
       <Partners />
+      <br />
       <Footer />
     </>
   );